refactor(server): extract leaf route domains — artifacts, safety, config, history (#585)
Split four fully-isolated read/write-leaf domains out of the AgentWireServer
god-class into routes/<domain>.py mixin modules, following the #584 seam:
ArtifactsRoutesMixin (3 handlers + /artifacts static mount), SafetyRoutesMixin
(4), ConfigRoutesMixin (3), HistoryRoutesMixin (3). Handlers moved verbatim;
inline add_* blocks in _setup_routes replaced by register_*_routes() calls;
mixins added to the AgentWireServer bases. Dropped now-unused security imports.

Pure relocation — route set invariant. Route-snapshot guard + full suite green.

Closes #585

Built by dotdev.dev
